moved repositories to a folder removed thymeleaf dependency
