+++ twitter-crawler/src/main/java/amcmurray/bw/twittercrawler/repositories/MentionRepository.java
@@ -1,4 +1,4 @@
-package amcmurray.bw.twittercrawler;
+package amcmurray.bw.twittercrawler.repositories;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
