+++ twitter-domain-objects/pom.xml
@@ -2,7 +2,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-
     <parent>
         <groupId>amcmurray.bw</groupId>
         <artifactId>twitter-crawler-parent</artifactId>
@@ -12,7 +11,6 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>twitter-domain-objects</artifactId>
-
 
     <dependencies>
         <dependency>
